Initial customization commit
This is a proof-of-concept. For anything intended for production, I'll
clean things up considerably.

The basic idea behind this plugin is to read RPC calls, inspect the method,
rewrite the path based on the method, and send the call back to traefik to
be routed to a service.

I'm working on a provider plugin to generate the config, but for now there's
a sample in the cfg/ directory.